backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
scipy
